+++ include/gtirb_pprinter/PrettyPrinter.hpp
@@ -348,6 +348,10 @@ protected:
                                   const gtirb::Section& section);
   virtual void printSectionFooterDirective(std::ostream& os,
                                            const gtirb::Section& section) = 0;
+
+  // Print the function ending of the function whose name is the given symbol.
+  virtual void printFunctionEnd(std::ostream& OS,
+                                const gtirb::Symbol& FunctionSymbol) = 0;
   virtual void printBlock(std::ostream& os, const gtirb::CodeBlock& block);
   virtual void printBlock(std::ostream& os, const gtirb::DataBlock& block);
   virtual void printBlockContents(std::ostream& os,
@@ -460,15 +464,19 @@ protected:
   gtirb::Context& context;
   const gtirb::Module& module;
 
-  std::optional<std::string> getContainerFunctionName(gtirb::Addr Addr) const;
-  virtual std::string getFunctionName(gtirb::Addr x) const;
+  [[deprecated(
+      "Use getContainerFunctionSymbol instead.")]] std::optional<std::string>
+  getContainerFunctionName(gtirb::Addr Addr) const;
+  [[deprecated]] virtual std::string getFunctionName(gtirb::Addr x) const;
+  [[deprecated("Use attribute FunctionFirstBlocks instead.")]] bool
+  isFunctionEntry(gtirb::Addr Addr) const;
+  [[deprecated("Use attribute FunctionLastBlocks instead.")]] bool
+  isFunctionLastBlock(gtirb::Addr Addr) const;
+
   virtual std::string getSymbolName(const gtirb::Symbol& symbol) const;
   virtual std::optional<std::string>
   getForwardedSymbolName(const gtirb::Symbol* symbol) const;
   virtual gtirb::Symbol* getForwardedSymbol(const gtirb::Symbol* Sym) const;
-
-  bool isFunctionEntry(gtirb::Addr Addr) const;
-  bool isFunctionLastBlock(gtirb::Addr Addr) const;
 
   // Currently, this only works for symbolic expressions in data blocks.
   // For the symbolic expressions that are part of code blocks, Capstone
@@ -506,9 +514,45 @@ private:
 
   static bool x86InstHasMoffsetEncoding(const cs_insn& inst);
 
+  /** Populate Function-related fields.*/
+  void computeFunctionInformation();
+  /** Populate AmbiguousSymbols */
+  void computeAmbiguousSymbols();
+
+  /** Get the symbol of the function that contains the block.
+   * This could return `nullptr` if the block does not belong to any function
+   * or if the function does not have any symbol associated to it.*/
+  const gtirb::Symbol*
+  getContainerFunctionSymbol(const gtirb::UUID& Uuid) const;
+
+  // A function is skipped if its name or any of its aliases are
+  // in the function skip policy.
+  bool isFunctionSkipped(const PrintingPolicy& Policy,
+                         const gtirb::Symbol& FunctionSymbol) const;
+
+  /** Mapping from function UUIDs to the symbols that define the function
+   * name.*/
+  std::map<gtirb::UUID, const gtirb::Symbol*> FunctionToSymbols;
+  /** Mapping from Block UUIDs to Function UUIDs.*/
+  std::map<gtirb::UUID, gtirb::UUID> BlockToFunction;
+  /** Set of blocks that are the first in each function.*/
+  std::set<gtirb::UUID> FunctionFirstBlocks;
+  /** Set of block UUIDS that are the last in each function.*/
+  std::set<gtirb::UUID> FunctionLastBlocks;
+
 protected:
-  std::set<gtirb::Addr> functionEntry;
-  std::set<gtirb::Addr> functionLastBlock;
+  [[deprecated("Use FunctionFirstBlocks instead.")]] std::set<gtirb::Addr>
+      functionEntry;
+  [[deprecated("Use FunctionLastBlocks instead.")]] std::set<gtirb::Addr>
+      functionLastBlock;
+
+  /** The set of all symbols associated to a function.*/
+  std::set<const gtirb::Symbol*> FunctionSymbols;
+  /** Mapping from function names to aliases. These are computed depending on
+   * the file format.*/
+  std::map<const gtirb::Symbol*, std::set<const gtirb::Symbol*>>
+      FunctionAliases;
+
   std::map<const gtirb::Symbol*, std::string> AmbiguousSymbols;
   std::string m_accum_comment;
   static std::string s_symaddr_0_warning(uint64_t symAddr);