+++ src/gtirb_pprinter/AttPrettyPrinter.cpp
@@ -101,9 +101,8 @@ void AttPrettyPrinter::printOpRegdirect(std::ostream& os, const cs_insn& inst,
   assert(op.type == X86_OP_REG &&
          "printOpRegdirect called without a register operand");
   if (cs_insn_group(this->csHandle, &inst, CS_GRP_CALL) ||
-      cs_insn_group(this->csHandle, &inst, CS_GRP_JUMP)) {
+      cs_insn_group(this->csHandle, &inst, CS_GRP_JUMP))
     os << '*';
-  }
   os << getRegisterName(op.reg);
 }
 
@@ -180,11 +179,10 @@ void AttPrettyPrinter::printOpIndirect(
   bool has_segment = op.mem.segment != X86_REG_INVALID;
   bool has_base = op.mem.base != X86_REG_INVALID;
   bool has_index = op.mem.index != X86_REG_INVALID;
-  
+
   if (cs_insn_group(this->csHandle, &inst, CS_GRP_CALL) ||
-      cs_insn_group(this->csHandle, &inst, CS_GRP_JUMP)) {
+      cs_insn_group(this->csHandle, &inst, CS_GRP_JUMP))
     os << '*';
-  }
 
   if (has_segment) {
     os << getRegisterName(op.mem.segment) << ':';