+++ src/gtirb_pprinter/AuxDataUtils.cpp
@@ -62,6 +62,10 @@ getFunctionBlocks(const gtirb::Module& Mod) {
   return util::getOrDefault<gtirb::schema::FunctionBlocks>(Mod);
 }
 
+std::map<gtirb::UUID, gtirb::UUID> getFunctionNames(const gtirb::Module& Mod) {
+  return util::getOrDefault<gtirb::schema::FunctionNames>(Mod);
+}
+
 std::optional<std::vector<CFIDirective>>
 getCFIDirectives(const gtirb::Offset& Offset, const gtirb::Module& Mod) {
   if (auto Lst = util::getByOffset<gtirb::schema::CfiDirectives>(Offset, Mod)) {
@@ -146,23 +150,20 @@ void setElfSymbolInfo(gtirb::Symbol& Sym, aux_data::ElfSymbolInfo& Info) {
   (*Table)[Sym.getUUID()] = Info.asAuxData();
 }
 
-bool hasVersionedSymDefs(const gtirb::IR& IR) {
-  for (const gtirb::Module& Module : IR.modules()) {
-    if (Module.getFileFormat() != gtirb::FileFormat::ELF) {
-      continue;
-    }
-
-    auto SymbolVersions = aux_data::getSymbolVersions(Module);
-    if (!SymbolVersions) {
-      continue;
-    }
-
-    auto& [SymVerDefs, SymVersNeeded, SymVerEntries] = *SymbolVersions;
-    if (SymVerDefs.size() > 0) {
-      return true;
-    }
+bool hasVersionedSymDefs(const gtirb::Module& Module) {
+  if (Module.getFileFormat() != gtirb::FileFormat::ELF) {
+    return false;
   }
 
+  auto SymbolVersions = aux_data::getSymbolVersions(Module);
+  if (!SymbolVersions) {
+    return false;
+  }
+
+  auto& [SymVerDefs, SymVersNeeded, SymVerEntries] = *SymbolVersions;
+  if (SymVerDefs.size() > 0) {
+    return true;
+  }
   return false;
 }