+++ src/gtirb_pprinter/MasmPrettyPrinter.cpp
@@ -493,6 +493,9 @@ std::string MasmPrettyPrinter::getRegisterName(unsigned int Reg) const {
 void MasmPrettyPrinter::printSymbolDefinition(std::ostream& Stream,
                                               const gtirb::Symbol& Symbol) {
   std::string Name = getSymbolName(Symbol);
+  // In MASM procedures can be exported by declaring "PROC EXPORT"
+  // Non-procedures (data) need to be declared "PUBLIC" AND
+  // be specified in the .def file.
   bool Exported = Exports.count(Symbol.getUUID()) > 0;
   if (Symbol.getReferent<gtirb::DataBlock>()) {
     if (Exported) {
@@ -503,17 +506,29 @@ void MasmPrettyPrinter::printSymbolDefinition(std::ostream& Stream,
     const gtirb::CodeBlock* Block = Symbol.getReferent<gtirb::CodeBlock>();
     bool SafeSeh = aux_data::getPeSafeExceptionHandlers(module).count(
                        Block->getUUID()) > 0;
-    if (Exported) {
-      Stream << Name << ' ' << masmSyntax.proc() << " EXPORT\n"
-             << Name << ' ' << masmSyntax.endp() << '\n';
-    } else if (SafeSeh) {
-      Stream << Name << ' ' << masmSyntax.proc() << "\n"
-             << ".SAFESEH " << Name << "\n"
-             << Name << ' ' << masmSyntax.endp() << '\n';
+    bool FunctionSymbol = FunctionSymbols.count(&Symbol) > 0;
+    if (FunctionSymbol) {
+      Stream << Name << ' ' << masmSyntax.proc();
+      if (Exported) {
+        Stream << " EXPORT";
+      }
+      Stream << "\n";
+      if (SafeSeh) {
+        Stream << ".SAFESEH " << Name << "\n";
+      }
     } else {
-      Stream << Name << ":\n";
+      if (Exported) {
+        Stream << syntax.global() << ' ' << Name << '\n';
+      }
+      // double colon makes labels available outside procedures
+      Stream << Name << "::\n";
     }
   }
+}
+
+void MasmPrettyPrinter::printFunctionEnd(std::ostream& OS,
+                                         const gtirb::Symbol& FunctionSymbol) {
+  OS << getSymbolName(FunctionSymbol) << ' ' << masmSyntax.endp() << '\n';
 }
 
 void MasmPrettyPrinter::printSymbolDefinitionRelativeToPC(