+++ src/gtirb_pprinter/Registration.cpp
@@ -32,6 +32,7 @@ void registerAuxDataTypes() {
   gtirb::AuxDataContainer::registerAuxDataType<Comments>();
   gtirb::AuxDataContainer::registerAuxDataType<FunctionEntries>();
   gtirb::AuxDataContainer::registerAuxDataType<FunctionBlocks>();
+  gtirb::AuxDataContainer::registerAuxDataType<FunctionNames>();
   gtirb::AuxDataContainer::registerAuxDataType<SymbolForwarding>();
   gtirb::AuxDataContainer::registerAuxDataType<Encodings>();
   gtirb::AuxDataContainer::registerAuxDataType<SectionProperties>();
@@ -53,6 +54,8 @@ void registerAuxDataTypes() {
   gtirb::AuxDataContainer::registerAuxDataType<ElfSymbolTabIdxInfo>();
   gtirb::AuxDataContainer::registerAuxDataType<ElfDynamicInit>();
   gtirb::AuxDataContainer::registerAuxDataType<ElfDynamicFini>();
+  gtirb::AuxDataContainer::registerAuxDataType<ElfStackExec>();
+  gtirb::AuxDataContainer::registerAuxDataType<ElfStackSize>();
 }
 
 void registerPrettyPrinters() {